Additional Fx60 fixes
Fx60 no longer honors the 'unpacked' flag in install.rdf for new
extension installs, so the Perl script used on macOS and Linux needs to
be copied out of the XPI to the Zotero temp directory before it can be
run.

I also started to fix the code for integration support before realizing
that Word 2016 and up no longer have a script menu, so I disabled that
part.